FIN DES 4 PARTES DE TIERS CLIENTS FOURN FOURNCLIENTS

+++ app/Http/Controllers/ProspectController.php
@@ -13,20 +13,23 @@ use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
 use Dompdf\Dompdf;
 use Dompdf\Options;
+use App\Models\Setting;
+
 
 class ProspectController extends Controller
 {
 
-    public function store (Request $request) {
+    public function store(Request $request)
+    {
         $rules = [
-            'nom_prospect' => ['nullable','max:50','string'],
-            'email_prospect' => ['nullable','email:rfc,dns','string','max:266','unique:prospects,email_prospect'],
-            'tele_prospect' => ['nullable','regex:/^\+?[0-9]{9,15}$/','unique:prospects,tele_prospect'],
-            'ville_prospect' => ['required','max:60','string'],
-            'nomSociete_prospect' => ['nullable','max:200','unique:prospects,nomSociete_prospect'],
-            'GSM1_prospect' => ['nullable','regex:/^\+?[0-9]{9,15}$/','unique:prospects,GSM1_prospect'],
-            'GSM2_prospect' => ['nullable','regex:/^\+?[0-9]{9,15}$/','unique:prospects,GSM2_prospect'],
-            'categorie_id' => ['required','integer','exists:categories,id'],
+            'nom_prospect' => ['nullable', 'max:50', 'string'],
+            'email_prospect' => ['nullable', 'email:rfc,dns', 'string', 'max:266', 'unique:prospects,email_prospect'],
+            'tele_prospect' => ['nullable', 'regex:/^\+?[0-9]{9,15}$/', 'unique:prospects,tele_prospect'],
+            'ville_prospect' => ['required', 'max:60', 'string'],
+            'nomSociete_prospect' => ['nullable', 'max:200', 'unique:prospects,nomSociete_prospect'],
+            'GSM1_prospect' => ['nullable', 'regex:/^\+?[0-9]{9,15}$/', 'unique:prospects,GSM1_prospect'],
+            'GSM2_prospect' => ['nullable', 'regex:/^\+?[0-9]{9,15}$/', 'unique:prospects,GSM2_prospect'],
+            'categorie_id' => ['required', 'integer', 'exists:categories,id'],
         ];
 
         $messages = [
@@ -50,19 +53,19 @@ class ProspectController extends Controller
             'categorie_id.exists' => 'Cette catégorie n\'existe pas!',
         ];
 
-        $validator = Validator::make($request->all(),$rules,$messages);
+        $validator = Validator::make($request->all(), $rules, $messages);
 
-        if ($validator->fails()){
+        if ($validator->fails()) {
             // dd($validator);
             return redirect()->back()
-                   ->withInput()
-                   ->with('modalType','default')
-                   ->withErrors($validator);
+                ->withInput()
+                ->with('modalType', 'default')
+                ->withErrors($validator);
         }
 
         $prospect = new Prospect();
         $prospect->nom_prospect = $request->nom_prospect ?? '';
-        $prospect->GSM1_prospect = $request->GSM1_prospect ?? '' ;
+        $prospect->GSM1_prospect = $request->GSM1_prospect ?? '';
         $prospect->GSM2_prospect = $request->GSM2_prospect ?? '';
         $prospect->ville_prospect = $request->ville_prospect;
         $prospect->tele_prospect = $request->tele_prospect ?? '';
@@ -75,71 +78,92 @@ class ProspectController extends Controller
         $categorie = Categorie::find($request->categorie_id);
         $categorie->prospects()->attach($prospect->id);
 
-        alert()->success('succès', $prospect->nom_prospect." ".'a été enregistrée avec succès.');
+        // Update the 'addedToday' in the settings table
+        $setting = Setting::where('key', 'tiersTracking')->first();
+
+        // If settings are not found, initialize them
+        if (!$setting) {
+            $setting = Setting::create([
+                'key' => 'tiersTracking',
+                'value' => 0,
+                'addedToday' => 0,
+                'deletedToday' => 0,
+            ]);
+        }
+
+        // Increment the 'addedToday' counter
+        $setting->increment('addedToday');
+
+        alert()->success('succès', $prospect->nom_prospect . " " . 'a été enregistrée avec succès.');
         return redirect()->to(url()->previous());
     }
 
     public function updateUserProspect(Request $request, $id)
-        {
+    {
 
-            $request->validate([
-                'user_id' => 'nullable|exists:users,id',
-            ]);
-
-
-            $prospect = Prospect::findOrFail($id);
-
-            $prospects = Prospect::where('groupId_prospect', $prospect->groupId_prospect)->get();
-
-            foreach ($prospects as $p) {
-                $p->user_id = $request->user_id ?? $p->user_id;
-                $p->save();
-            }
+        $request->validate([
+            'user_id' => 'nullable|exists:users,id',
+        ]);
 
 
-            return redirect()->back();        }
+        $prospect = Prospect::findOrFail($id);
 
-            public function updateRemarkProspect(Request $request, $id)
-            {
-                $validator = Validator::make($request->all(), [
-                    'remark' => ['nullable','string',function ($attribute, $value, $fail) {
-                        $wordCount = str_word_count($value);
-                        if ($wordCount > 100) {
-                            $fail('La description ne doit pas dépasser 100 mots.');
-                        }
-                    }]
+        $prospects = Prospect::where('groupId_prospect', $prospect->groupId_prospect)->get();
 
-                ]
-            
-                ,[
-                    "remark.string" => "La remarque doit etre de type chaine de caractere" 
-                ]);
-
-                if ($validator->fails()) {
-                    return redirect()->back()
-                    ->withInput()
-                    ->with('modalType', 'remark')
-                    ->withErrors($validator);
-                }
-
-                $prospect = Prospect::findOrFail($id);
-
-                $prospects = Prospect::where('groupId_prospect', $prospect->groupId_prospect)->get();
-
-                foreach ($prospects as $p) {
-                    $p->remark = $request->remark ;
-                    $p->save();
-                }
-
-                return redirect()->back();        }
+        foreach ($prospects as $p) {
+            $p->user_id = $request->user_id ?? $p->user_id;
+            $p->save();
+        }
 
 
+        return redirect()->back();
+    }
+
+    public function updateRemarkProspect(Request $request, $id)
+    {
+        $validator = Validator::make(
+            $request->all(),
+            [
+                'remark' => ['nullable', 'string', function ($attribute, $value, $fail) {
+                    $wordCount = str_word_count($value);
+                    if ($wordCount > 100) {
+                        $fail('La description ne doit pas dépasser 100 mots.');
+                    }
+                }]
+
+            ],
+            [
+                "remark.string" => "La remarque doit etre de type chaine de caractere"
+            ]
+        );
+
+        if ($validator->fails()) {
+            return redirect()->back()
+                ->withInput()
+                ->with('modalType', 'remark')
+                ->withErrors($validator);
+        }
+
+        $prospect = Prospect::findOrFail($id);
+
+        $prospects = Prospect::where('groupId_prospect', $prospect->groupId_prospect)->get();
+
+        foreach ($prospects as $p) {
+            $p->remark = $request->remark;
+            $p->save();
+        }
+
+        return redirect()->back();
+    }
 
 
-    public function index (Request $request) {
 
-        $perPage = $request->get('per_page',10);
-        $prospects = Prospect::with('categories','categorieProspects.categorie','utilisateur')->paginate($perPage);
+
+    public function index(Request $request)
+    {
+
+        $perPage = $request->get('per_page', 10);
+        $prospects = Prospect::with('categories', 'categorieProspects.categorie', 'utilisateur')->paginate($perPage);
 
         $categories = Categorie::with('sousCategories')->get();
 
@@ -149,55 +173,78 @@ class ProspectController extends Controller
 
         $select = ['Fournisseur', 'Client', 'Client et Fournisseur'];
 
-        return view('myApp.admin.links.prospects',compact('prospects','categories','select','perPage'));
+        return view('myApp.admin.links.prospects', compact('prospects', 'categories', 'select', 'perPage'));
     }
 
     public function prospectsPdf()
     {
-            $prospects = Prospect::with('categorieProspects.categorie')->get();
+        $prospects = Prospect::with('categorieProspects.categorie')->get();
 
-            $options = new Options();
-            $options->set('defaultFont', 'Courier');
-            $dompdf = new Dompdf($options);
+        $options = new Options();
+        $options->set('defaultFont', 'Courier');
+        $dompdf = new Dompdf($options);
 
-            $html = view('myApp/admin/pdf/prospects', compact('prospects'))->render();
+        $html = view('myApp/admin/pdf/prospects', compact('prospects'))->render();
 
-            $dompdf->loadHtml($html);
+        $dompdf->loadHtml($html);
 
-            $dompdf->setPaper('A4', 'portrait');
+        $dompdf->setPaper('A4', 'portrait');
 
-            $dompdf->render();
+        $dompdf->render();
 
-            return response($dompdf->output(), 200)
-                ->header('Content-Type', 'application/pdf')
-                ->header('Content-Disposition', 'attachment; filename="prospects-list.pdf"');
+        return response($dompdf->output(), 200)
+            ->header('Content-Type', 'application/pdf')
+            ->header('Content-Disposition', 'attachment; filename="prospects-list.pdf"');
     }
 
 
 
 
-    public function destroy($id){
+    public function destroy($id)
+    {
         $prospect = Prospect::find($id);
-        $prospect->delete();
+
+        if ($prospect) {
+            $prospect->delete();
+
+            // Update the 'deletedToday' in the settings table
+            $setting = Setting::where('key', 'tiersTracking')->first();
+
+            // If settings are not found, initialize them
+            if (!$setting) {
+                $setting = Setting::create([
+                    'key' => 'tiersTracking',
+                    'value' => 0,
+                    'addedToday' => 0,
+                    'deletedToday' => 0,
+                ]);
+            }
+
+            // Increment the 'deletedToday' counter
+            $setting->increment('deletedToday');
+        }
+
         return redirect()->to(url()->previous());
     }
 
 
 
 
-    public function update (Request $request) {
+
+    public function update(Request $request)
+    {
 
         $prospect = Prospect::find((int)$request->id);
 
         $rules = [
-            'newNom_prospect' => ['nullable','max:50','string'],
-            'newEmail_prospect' => ['nullable','email:rfc,dns','string','max:266'],
-            'newTele_prospect' => ['nullable','regex:/^\+?[0-9]{9,15}$/'],
-            'newVille_prospect' => ['required','max:60','string'],
-            'newNomSociete_prospect' => ['nullable','max:200'],
-            'newGSM1_prospect' => ['nullable','regex:/^\+?[0-9]{9,15}$/'],
-            'newGSM2_prospect' => ['nullable','regex:/^\+?[0-9]{9,15}$/'],
-            'newCategorie_id' => ['required','integer','exists:categories,id'],
+            'newNom_prospect' => ['nullable', 'max:50', 'string'],
+            'newEmail_prospect' => ['nullable', 'email:rfc,dns', 'string', 'max:266'],
+            'newTele_prospect' => ['nullable', 'regex:/^\+?[0-9]{9,15}$/'],
+            'newVille_prospect' => ['required', 'max:60', 'string'],
+            'newNomSociete_prospect' => ['nullable', 'max:200'],
+            'newGSM1_prospect' => ['nullable', 'regex:/^\+?[0-9]{9,15}$/'],
+            'newGSM2_prospect' => ['nullable', 'regex:/^\+?[0-9]{9,15}$/'],
+            'newCategorie_id' => ['required', 'integer', 'exists:categories,id'],
         ];
 
         $messages = [
@@ -216,13 +263,13 @@ class ProspectController extends Controller
             'newCategorie_id.exists' => 'Cette catégorie n\'existe pas!',
         ];
 
-        $validator = Validator::make($request->all(),$rules,$messages);
+        $validator = Validator::make($request->all(), $rules, $messages);
 
-        if ($validator->fails()){
+        if ($validator->fails()) {
             return redirect()->back()
-                   ->withInput()
-                   ->with('modalType','update')
-                   ->withErrors($validator);
+                ->withInput()
+                ->with('modalType', 'update')
+                ->withErrors($validator);
         }
 
         $nomSociety = $request->newNomSociete_prospect ?? '';
@@ -232,18 +279,17 @@ class ProspectController extends Controller
         $GSM1 = $request->newGSM1_prospect ?? '';
         $GSM2 = $request->newGSM2_prospect ?? '';
         $newCategorieId = $request->newCategorie_id;
-        $existingProspect = Prospect::
-        where('nom_prospect', $name)
-        ->where('email_prospect', $email)
-        ->where('tele_prospect', $tele)
-        ->where('ville_prospect', $request->newVille_prospect)
-        ->where('nomSociete_prospect', $nomSociety)
-        ->where('GSM1_prospect', $GSM1)
-        ->where('GSM2_prospect', $GSM2)
-        ->whereHas('categories', function ($query) use ($newCategorieId) {
-            $query->where('categories.id', $newCategorieId);
-        })
-        ->first();
+        $existingProspect = Prospect::where('nom_prospect', $name)
+            ->where('email_prospect', $email)
+            ->where('tele_prospect', $tele)
+            ->where('ville_prospect', $request->newVille_prospect)
+            ->where('nomSociete_prospect', $nomSociety)
+            ->where('GSM1_prospect', $GSM1)
+            ->where('GSM2_prospect', $GSM2)
+            ->whereHas('categories', function ($query) use ($newCategorieId) {
+                $query->where('categories.id', $newCategorieId);
+            })
+            ->first();
 
         if ($existingProspect) {
             alert()->error('Erreur', 'Une version avec les mêmes informations existe déjà. Veuillez modifier au moins un champ.');
@@ -275,34 +321,33 @@ class ProspectController extends Controller
         }
 
         $prospectsSimilaires = Prospect::where('groupId_prospect', $prospect->groupId_prospect)
-        ->where('id', '!=', $prospect->id)
-        ->get();
+            ->where('id', '!=', $prospect->id)
+            ->get();
 
-       foreach ($prospectsSimilaires as $similarProspect) {
-        $similarProspect->nom_prospect = $request->newNom_prospect ?? '';
-        $similarProspect->email_prospect = $request->newEmail_prospect ?? '';
-        $similarProspect->tele_prospect = $request->newTele_prospect ?? '';
-        $similarProspect->ville_prospect = $request->newVille_prospect;
-        $similarProspect->nomSociete_prospect = $request->newNomSociete_prospect ?? '';
-        $similarProspect->GSM1_prospect = $request->newGSM1_prospect ?? '';
-        $similarProspect->GSM2_prospect = $request->newGSM2_prospect ?? '';
-        if ($similarProspect->save()) {
-            alert()->success('Succès', 'Le fournisseur a été mis à jour avec succès.');
+        foreach ($prospectsSimilaires as $similarProspect) {
+            $similarProspect->nom_prospect = $request->newNom_prospect ?? '';
+            $similarProspect->email_prospect = $request->newEmail_prospect ?? '';
+            $similarProspect->tele_prospect = $request->newTele_prospect ?? '';
+            $similarProspect->ville_prospect = $request->newVille_prospect;
+            $similarProspect->nomSociete_prospect = $request->newNomSociete_prospect ?? '';
+            $similarProspect->GSM1_prospect = $request->newGSM1_prospect ?? '';
+            $similarProspect->GSM2_prospect = $request->newGSM2_prospect ?? '';
+            if ($similarProspect->save()) {
+                alert()->success('Succès', 'Le fournisseur a été mis à jour avec succès.');
+            };
         }
-
-        ;
-    }
 
         return redirect()->back();
     }
 
-    public function search(Request $request){
+    public function search(Request $request)
+    {
         $select = ['Fournisseur', 'Client', 'Client et Fournisseur'];
         $search = $request->input('search');
-        $prospect = Prospect :: with(['categories.sousCategories','utilisateur'])
-        ->where('nom_prospect','LIKE',"%{$search}%")
-        ->orWhere('nomSociete_prospect','LIKE',"%{$search}%")
-        ->get();
+        $prospect = Prospect::with(['categories.sousCategories', 'utilisateur'])
+            ->where('nom_prospect', 'LIKE', "%{$search}%")
+            ->orWhere('nomSociete_prospect', 'LIKE', "%{$search}%")
+            ->get();
         return response()->json([
             'prospects' => $prospect,
             'selectOptions' => $select
@@ -313,13 +358,13 @@ class ProspectController extends Controller
     private function hasOtherChanges($prospect, $request)
     {
 
-        return $prospect->nom_prospect !== ($request->newNom_prospect ?? '')||
-               $prospect->email_prospect !== ($request->newEmail_prospect ?? '')||
-               $prospect->tele_prospect !== ($request->newTele_prospect ?? '')||
-               $prospect->ville_prospect !== $request->newVille_prospect||
-               $prospect->nomSociete_prospect !== ($request->newNomSociete_prospect ?? '')||
-               $prospect->GSM1_prospect !== ($request->newGSM1_prospect ?? '') ||
-               $prospect->GSM2_prospect !== ($request->newGSM2_prospect ?? '');
+        return $prospect->nom_prospect !== ($request->newNom_prospect ?? '') ||
+            $prospect->email_prospect !== ($request->newEmail_prospect ?? '') ||
+            $prospect->tele_prospect !== ($request->newTele_prospect ?? '') ||
+            $prospect->ville_prospect !== $request->newVille_prospect ||
+            $prospect->nomSociete_prospect !== ($request->newNomSociete_prospect ?? '') ||
+            $prospect->GSM1_prospect !== ($request->newGSM1_prospect ?? '') ||
+            $prospect->GSM2_prospect !== ($request->newGSM2_prospect ?? '');
     }
 
     public function prospect(Request $request, $id)
@@ -356,8 +401,7 @@ class ProspectController extends Controller
                 }
                 $prospectItem->delete();
             }
-
-        }else if ($selectedStatus === 'Client') {
+        } else if ($selectedStatus === 'Client') {
 
             foreach ($prospectsGroup as $prospectItem) {
 
@@ -384,9 +428,8 @@ class ProspectController extends Controller
 
 
                 $prospectItem->delete();
-
             }
-        }else if ($selectedStatus === 'Client et Fournisseur') {
+        } else if ($selectedStatus === 'Client et Fournisseur') {
             foreach ($prospectsGroup as $prospectItem) {
 
                 $fc = new FournisseurClient();
@@ -412,9 +455,8 @@ class ProspectController extends Controller
 
 
                 $prospectItem->delete();
-
             }
         }
-     return redirect()->to(url()->previous());
-}
+        return redirect()->to(url()->previous());
+    }
 }
